+++ lib/alignment.dart
@@ -2,13 +2,13 @@ import 'package:dungeon_world_data/_base.dart';
 
 class Alignment extends DWEntity {
   /// Alignment key
-  final String key;
+  String key;
 
   /// Alignment name
-  final String name;
+  String name;
 
   /// Alignment description
-  final String description;
+  String description;
 
   Alignment(this.key, this.name, this.description);